moved array triangles into package and added more abstract class
