+++ autoarray/structures/triangles/array/__init__.py
@@ -1,0 +1,2 @@
+from .array import ArrayTriangles
+from .jax_array import ArrayTriangles as JAXArrayTriangles
