+++ autoarray/structures/triangles/jax_coordinate_array.py
@@ -6,7 +6,7 @@ from autoarray.structures.triangles.abstract import HEIGHT_FACTOR
 from autoarray.structures.triangles.abstract_coordinate_array import (
     AbstractCoordinateArray,
 )
-from autoarray.structures.triangles.jax_array import ArrayTriangles
+from autoarray.structures.triangles.array.jax_array import ArrayTriangles
 from autoarray.numpy_wrapper import register_pytree_node_class
 from autoconf import cached_property
