+++ test_autoarray/structures/triangles/test_nan_triangles.py
@@ -2,7 +2,7 @@ import pytest
 
 try:
     from jax import numpy as np
-    from autoarray.structures.triangles.jax_array import ArrayTriangles
+    from autoarray.structures.triangles.array.jax_array import ArrayTriangles
 except ImportError:
     import numpy as np
     from autoarray.structures.triangles.array import ArrayTriangles
